requirements.txt
numpy==1.19.2
PyOpenGL==3.1.5
PyQt5==5.14.0
